requirements.txt
#Add the package name to your Astro project. 
#If it’s a Python package, add it to requirements.txt. 
#If it’s an OS-level package, add it to packages.txt. 
#The latest version of the package that’s publicly available is installed by default.

#numpy==1.23.0
apache-airflow-providers-common-sql
apache-airflow-providers-snowflake
snowflake-connector-python
snowflake-sqlalchemy
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
boto3
botocore